requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==8.0.3
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
dataclasses==0.8
Flask==2.0.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==4.8.2
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
numpy==1.19.5
oauthlib==2.0.6
opencv-python==4.4.0.46
packaging==21.3
Pillow==8.3.1
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
scikit-build==0.12.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
Twisted==17.9.0
typing_extensions==4.0.1
urllib3==1.22
Werkzeug==2.0.2
zipp==3.6.0
